Improved compatibility with JS used by Magento
In current versions of Magento (Community Edition, at least), Prototype messes
with the Array prototype. As a result, the part of SVGInjector that copies
attributes of the original `img` tag to the newly embedded SVG `tag`.

The fix is manually iterating over the `img` tag's attributes instead of using
`Array.filter()`, and adding the attributes matching the pattern to `imgData`.